+++ src/lexer/src/span.rs
@@ -1,8 +1,11 @@
 use std::fmt;
 
+/// Represents a line/column position in a script
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub struct Position {
+    /// The position's line, counting from 1
     pub line: u32,
+    /// The position's column, counting from 1.
     pub column: u32,
 }
 
@@ -18,8 +21,11 @@ impl fmt::Display for Position {
     }
 }
 
+/// A span is a range in the source code, represented by a start and end position
 #[derive(Clone, Copy, Debug, Default, PartialEq)]
 pub struct Span {
+    /// The span's start position
     pub start: Position,
+    /// The span's end position
     pub end: Position,
 }