+++ src/parser/src/parser.rs
@@ -209,6 +209,7 @@ impl ExpressionContext {
     }
 }
 
+/// Koto's parser
 pub struct Parser<'source> {
     ast: Ast,
     constants: ConstantPoolBuilder,
@@ -217,6 +218,7 @@ pub struct Parser<'source> {
 }
 
 impl<'source> Parser<'source> {
+    /// Takes in a source script, and produces an Ast and associated [ConstantPool]
     pub fn parse(source: &'source str) -> Result<(Ast, ConstantPool), ParserError> {
         let capacity_guess = source.len() / 4;
         let mut parser = Parser {
@@ -503,7 +505,7 @@ impl<'source> Parser<'source> {
         } else if let Some(result) = self.parse_export(context)? {
             Some(result)
         } else {
-            self.parse_expressions(context, false)?
+            self.parse_expressions(context, TempResult::No)?
         };
 
         self.frame_mut()?.finish_expression();
@@ -511,10 +513,18 @@ impl<'source> Parser<'source> {
         Ok(result)
     }
 
+    // Parse a comma separated series of expressions
+    //
+    // If only a single expression is encountered then that expression's node is the only thing
+    // generated.
+    //
+    // Otherwise, for multiple expressions, the result of the expression can be temporary
+    // (i.e. not assigned to an identifier) in which case a TempTuple is generated,
+    // otherwise the result will be a Tuple.
     fn parse_expressions(
         &mut self,
         context: &mut ExpressionContext,
-        temp_result: bool,
+        temp_result: TempResult,
     ) -> Result<Option<AstIndex>, ParserError> {
         let mut expression_context = ExpressionContext {
             allow_space_separated_call: true,
@@ -556,10 +566,9 @@ impl<'source> Parser<'source> {
             if expressions.len() == 1 && !encountered_comma {
                 Ok(Some(first))
             } else {
-                let result = if temp_result {
-                    Node::TempTuple(expressions)
-                } else {
-                    Node::Tuple(expressions)
+                let result = match temp_result {
+                    TempResult::No => Node::Tuple(expressions),
+                    TempResult::Yes => Node::TempTuple(expressions),
                 };
                 Ok(Some(self.push_node(result)?))
             }
@@ -721,8 +730,15 @@ impl<'source> Parser<'source> {
         }
 
         let single_target = targets.len() == 1;
+
+        let temp_result = if single_target {
+            TempResult::No
+        } else {
+            TempResult::Yes
+        };
+
         if let Some(rhs) =
-            self.parse_expressions(&mut ExpressionContext::permissive(), !single_target)?
+            self.parse_expressions(&mut ExpressionContext::permissive(), temp_result)?
         {
             let node = if single_target {
                 Node::Assign {
@@ -1265,7 +1281,7 @@ impl<'source> Parser<'source> {
                 self.consume_next_token_on_same_line();
 
                 if let Some(rhs) =
-                    self.parse_expressions(&mut ExpressionContext::permissive(), false)?
+                    self.parse_expressions(&mut ExpressionContext::permissive(), TempResult::No)?
                 {
                     let node = Node::Assign {
                         target: AssignTarget {
@@ -1320,11 +1336,12 @@ impl<'source> Parser<'source> {
 
         let mut context = ExpressionContext::permissive();
         let expression_source_start = self.lexer.source_position();
-        let expression = if let Some(expression) = self.parse_expressions(&mut context, false)? {
-            expression
-        } else {
-            return syntax_error!(ExpectedExpression, self);
-        };
+        let expression =
+            if let Some(expression) = self.parse_expressions(&mut context, TempResult::No)? {
+                expression
+            } else {
+                return syntax_error!(ExpectedExpression, self);
+            };
 
         let expression_source_end = self.lexer.source_position();
 
@@ -1460,7 +1477,7 @@ impl<'source> Parser<'source> {
                 Token::Yield => {
                     self.consume_next_token(context);
                     if let Some(expression) =
-                        self.parse_expressions(&mut context.start_new_expression(), false)?
+                        self.parse_expressions(&mut context.start_new_expression(), TempResult::No)?
                     {
                         let result = self.push_node(Node::Yield(expression))?;
                         self.frame_mut()?.contains_yield = true;
@@ -1479,14 +1496,9 @@ impl<'source> Parser<'source> {
                 }
                 Token::Return => {
                     self.consume_next_token(context);
-                    let result = if let Some(expression) =
-                        self.parse_expressions(&mut context.start_new_expression(), false)?
-                    {
-                        self.push_node(Node::ReturnExpression(expression))?
-                    } else {
-                        self.push_node(Node::Return)?
-                    };
-                    Some(result)
+                    let return_value = self
+                        .parse_expressions(&mut context.start_new_expression(), TempResult::No)?;
+                    Some(self.push_node(Node::Return(return_value))?)
                 }
                 Token::Throw => self.parse_throw_expression()?,
                 Token::Debug => self.parse_debug_expression()?,
@@ -1818,15 +1830,19 @@ impl<'source> Parser<'source> {
             return syntax_error!(ExpectedForArgs, self);
         }
 
-        let range = match self.parse_expression(&mut ExpressionContext::inline())? {
-            Some(range) => range,
-            None => return syntax_error!(ExpectedForRanges, self),
+        let iterable = match self.parse_expression(&mut ExpressionContext::inline())? {
+            Some(iterable) => iterable,
+            None => return syntax_error!(ExpectedForIterable, self),
         };
 
         match self.parse_indented_block()? {
             Some(body) => {
                 let result = self.push_node_with_start_span(
-                    Node::For(AstFor { args, range, body }),
+                    Node::For(AstFor {
+                        args,
+                        iterable,
+                        body,
+                    }),
                     start_span,
                 )?;
 
@@ -1915,13 +1931,14 @@ impl<'source> Parser<'source> {
 
         let result = if self.peek_next_token_on_same_line() == Some(Token::Then) {
             self.consume_next_token_on_same_line();
-            let then_node = match self.parse_expressions(&mut ExpressionContext::inline(), true)? {
-                Some(then_node) => then_node,
-                None => return syntax_error!(ExpectedThenExpression, self),
-            };
+            let then_node =
+                match self.parse_expressions(&mut ExpressionContext::inline(), TempResult::No)? {
+                    Some(then_node) => then_node,
+                    None => return syntax_error!(ExpectedThenExpression, self),
+                };
             let else_node = if self.peek_next_token_on_same_line() == Some(Token::Else) {
                 self.consume_next_token_on_same_line();
-                match self.parse_expressions(&mut ExpressionContext::inline(), true)? {
+                match self.parse_expressions(&mut ExpressionContext::inline(), TempResult::No)? {
                     Some(else_node) => Some(else_node),
                     None => return syntax_error!(ExpectedElseExpression, self),
                 }
@@ -2029,7 +2046,7 @@ impl<'source> Parser<'source> {
                     self.consume_next_token_on_same_line();
 
                     if let Some(expression) =
-                        self.parse_expressions(&mut ExpressionContext::inline(), true)?
+                        self.parse_expressions(&mut ExpressionContext::inline(), TempResult::No)?
                     {
                         expression
                     } else if let Some(indented_expression) = self.parse_indented_block()? {
@@ -2040,7 +2057,9 @@ impl<'source> Parser<'source> {
                 }
                 Some(Token::Then) => {
                     self.consume_next_token_on_same_line();
-                    match self.parse_expressions(&mut ExpressionContext::inline(), true)? {
+                    match self
+                        .parse_expressions(&mut ExpressionContext::inline(), TempResult::No)?
+                    {
                         Some(expression) => expression,
                         None => return syntax_error!(ExpectedSwitchArmExpressionAfterThen, self),
                     }
@@ -2096,7 +2115,7 @@ impl<'source> Parser<'source> {
         let start_span = self.current_span();
 
         let match_expression =
-            match self.parse_expressions(&mut ExpressionContext::inline(), true)? {
+            match self.parse_expressions(&mut ExpressionContext::inline(), TempResult::Yes)? {
                 Some(expression) => expression,
                 None => {
                     return syntax_error!(ExpectedMatchExpression, self);
@@ -2166,7 +2185,7 @@ impl<'source> Parser<'source> {
                     self.consume_next_token_on_same_line();
 
                     if let Some(expression) =
-                        self.parse_expressions(&mut ExpressionContext::inline(), true)?
+                        self.parse_expressions(&mut ExpressionContext::inline(), TempResult::No)?
                     {
                         expression
                     } else if let Some(indented_expression) = self.parse_indented_block()? {
@@ -2181,7 +2200,9 @@ impl<'source> Parser<'source> {
                     }
 
                     self.consume_next_token_on_same_line();
-                    match self.parse_expressions(&mut ExpressionContext::inline(), true)? {
+                    match self
+                        .parse_expressions(&mut ExpressionContext::inline(), TempResult::No)?
+                    {
                         Some(expression) => expression,
                         None => return syntax_error!(ExpectedMatchArmExpressionAfterThen, self),
                     }
@@ -2370,10 +2391,10 @@ impl<'source> Parser<'source> {
         // Mark any imported ids as locally assigned
         for item in items.iter() {
             match item.last() {
-                Some(ImportItem::Id(id)) => {
+                Some(ImportItemNode::Id(id)) => {
                     self.frame_mut()?.ids_assigned_in_scope.insert(*id);
                 }
-                Some(ImportItem::Str(_)) => {}
+                Some(ImportItemNode::Str(_)) => {}
                 None => return internal_error!(ExpectedIdInImportItem, self),
             };
         }
@@ -2455,15 +2476,15 @@ impl<'source> Parser<'source> {
         Ok(Some(result))
     }
 
-    fn consume_import_items(&mut self) -> Result<Vec<Vec<ImportItem>>, ParserError> {
+    fn consume_import_items(&mut self) -> Result<Vec<Vec<ImportItemNode>>, ParserError> {
         let mut items = vec![];
         let mut item_context = ExpressionContext::permissive();
 
         loop {
             let item_root = match self.parse_id(&mut item_context)? {
-                Some(id) => ImportItem::Id(id),
+                Some(id) => ImportItemNode::Id(id),
                 None => match self.parse_string(&mut item_context)? {
-                    Some((import_string, _span)) => ImportItem::Str(import_string),
+                    Some((import_string, _span)) => ImportItemNode::Str(import_string),
                     None => break,
                 },
             };
@@ -2474,10 +2495,10 @@ impl<'source> Parser<'source> {
                 self.consume_token();
 
                 match self.parse_id(&mut ExpressionContext::restricted())? {
-                    Some(id) => item.push(ImportItem::Id(id)),
+                    Some(id) => item.push(ImportItemNode::Id(id)),
                     None => match self.parse_string(&mut ExpressionContext::restricted())? {
                         Some((node_string, _span)) => {
-                            item.push(ImportItem::Str(node_string));
+                            item.push(ImportItemNode::Str(node_string));
                         }
                         None => return syntax_error!(ExpectedImportModuleId, self),
                     },
@@ -2749,8 +2770,8 @@ impl<'source> Parser<'source> {
                     Some(CurlyOpen) => {
                         self.consume_token();
 
-                        if let Some(expression) =
-                            self.parse_expressions(&mut ExpressionContext::inline(), true)?
+                        if let Some(expression) = self
+                            .parse_expressions(&mut ExpressionContext::inline(), TempResult::No)?
                         {
                             nodes.push(StringNode::Expr(expression));
                         } else {
@@ -3057,6 +3078,11 @@ impl<'source> Parser<'source> {
 
         None
     }
+}
+
+enum TempResult {
+    No,
+    Yes,
 }
 
 // The first operator that's above the pipe operator >> in precedence.